requirements.txt
djangorestframework_simplejwt==4.3.0
Django==2.2.4
python-dotenv==0.10.3
djangorestframework==3.10.2
gunicorn==19.9.0
psycopg2==2.8.3
django-heroku==0.3.1
